Do not update global `python.defaultInterpreterPath` when a runtime starts (#5054)
Addresses #4552 

I want to point out #3255 where we _added_ updating the path in the
workspace. I suspect (???) that updating the setting globally was never
the right thing to do; this was added in the massive
https://github.com/posit-dev/positron-python/pull/413 for multi-session
support.

This PR does not address how updating the
`python.defaultInterpreterPath` user setting starts Python. I have [some
more notes
here](https://github.com/posit-dev/positron/issues/4552#issuecomment-2418322327)
and we can open a followup PR if folks think this is a problem. It seems
mostly like intended/inherited behavior from upstream to me.

### QA Notes

- Open a workspace and have R only running.
- Open an additional workspace and have R only running.
- Start a Python runtime in the second workspace via the top bar.

You should _not_ see any Python runtime start in the first
workspace/window.

We should still see the correct behavior for #3169 and #3255.

I don't think we can write a useful unit test for this and we don't
currently have multi-workspace tests. It might be worth putting that
(smoke tests for multiple workspaces) on the todo list eventually but
certainly seems like an advanced move.