Enable kernel supervisor download in Github workflows (#5053)
Enables downloads for the kernel supervisor to prepare for running tests
against it. Most of the work here is generating a temporary PAT to use
to download the supervisor.

It is somewhat duplicative to generate the PAT outside the composite
action and pass it in every time, but composite actions don't support
secrets (unlike reusable workflows) and it feels safer to push an
ephemeral PAT into here than all the data we need to actually generate
one.