use calculated build number
The `positronBuildNumber` from the package.json file will always be zero -- instead, we need to use the calculated build number from show-version.js when possible.

We are already using the calculated build number in a few places, but a few were missing.

The `positronBuildVersion` variable is now in a separate util file, so the value can be used across the gulpfiles.