e2e: fix `notebookVariables.test.ts` (#5142)
### Intent

Stabilize the `notebookVariables.test.ts` test to prevent flakiness.

### Approach

I observed that this test was flaky, and when debugging, I couldn’t get
the second test to pass without restarting the app (same as the test was
doing). This indicated something was off. To address this, I replaced
the use of the quick access menu for editing the cell with a direct
manual selection. With this change, both tests now run successfully in
the same app instance without any issues.

### QA Notes
Passed on both Electron and Web.