Data Explorer: Compute column profiles in smaller batches for large data frames, avoid timeouts (#5148)
Addresses #4629 and #2851. This fixes timeouts that were occurring,
making it seem like the data explorer is broken on large datasets, while
improving the performance / responsiveness by computing updates
in small batches rather than in a monolithic batch which may take 10s of
seconds to return.

I found that if we spam the backend with all of the profile requests at
once, the `get_data_values` request doesn't get served because of all
the thread contention, so this just has one profile request active a
time.

You can see what a 33M row dataset looks like on the initial open and
then various filtering workflows


https://github.com/user-attachments/assets/9df973a7-9cc4-4237-806d-01dbf2872ec8

This does make the lack of a loading indicator somewhat more pronounced,
but that can be addressed separately.