Data Explorer: Implement row filters and column sorting for duckdb backend (#5168)
Addresses #5126 and #5127. It provides the full set of row filter types
and column sorting. A lot of the work was in writing the unit tests code
(with `yarn test-extension -l positron-duckdb`) to verify correctness,
but I was fairly thorough and turned up some bugs while writing the unit
tests.

DuckDB does not provide stable sorting (i.e. rows may be permuted if the
sort keys are equal) so I had to introduce an auxiliary "row_index"
field to do an order-preserving sort to make it easier to unit test.