Data Explorer: Do not send backend empty schema, data, profile requests (#5170)
I noticed that the data explorer UI was sending the backends empty
schema, data, and profile requests (length-0 array of requests), which
makes it look like computations are happening (the loading indicator
activates when scrolling) even though the data is fully cached. It is
also extra Jupyter comm chatter that is not needed.