Check if Python has sqlite3 is installed before attempting to install ipykernel (#5149)
Addresses #4698

Currently the message says this if your Python doesn't have sqlite3
there:

![Screenshot 2024-10-23 at 4 08
08 PM](https://github.com/user-attachments/assets/29a85fc9-4d2e-4101-a79c-49afc9cc2e1a)

Thoughts or feedback? We believe this is most likely to happen on Linux
when you are missing a sysreq.

### QA Notes

To confirm this is working as expected, you need two Python envs:

- Make one new Python env that is "normal", i.e. has the built-in
sqlite3 module, but doesn't have ipykernel installed (perhaps via `pyenv
virtualenv`). You should be able to choose this env via "Python: Select
Interpreter", agree to the ipykernel installation, and see the ipykernel
installation succeed. ✅
- Make one new Python env that has ✨ problems ✨ i.e. does not have the
built-in sqlite3 module. The only way I could find to do this is to
delete it: first [find where it is for the Python you are about to
destroy](https://stackoverflow.com/questions/2501746/deleting-python-modules)
and then delete it, i.e. like `rm
/Users/juliasilge/.pyenv/versions/3.11.5/lib/python3.11/lib-dynload/_sqlite3.cpython-311-darwin.so`.
Now choose this Python env via "Python: Select Interpreter" and see the
message telling you that you need sqlite3. 🚫