Make the Jupyter supervisor the default in non-desktop Positron configurations (#5215)
This change makes Kallichore the default in non-desktop configurations
of Positron, such as on Posit Workbench. In addition to the
configuration change, it includes a number of other changes that improve
the stability of the supervisor integration in order to bring it up to
the quality needed.

In particular:
- The supervisor's behavior around session starting has changed; the RPC
that starts sessions now doesn't return until the session has fully
started and returned kernel information. We pick up this kernel info in
Positron rather than making our own RPC for kernel info (when starting
for the first time).
- We now hold any attempts to make UI Comm requests until the UI Comm
has opened. This eliminates a lot of `setConsoleWidth` errors in the
logs.
- The Kallichore terminal is now hidden by default.
- The R and Python extensions no longer have a hard dependency on the
Jupyter Adapter extension. This allows them to work even when the
Jupyter Adapter is unbootable (such as on REHL 9 or other systems where
there are glibc issues with ZeroMQ)

Addresses #4941. 

### QA Notes

If logging issues, it's helpful to include the Kallichore logs, so run
with the "Show Terminal" setting (added in this PR) turned on.

---------

Signed-off-by: Jonathan <jonathan@posit.co>
Co-authored-by: sharon <sharon-wang@users.noreply.github.com>