Data Explorer: Display duckdb error message when file fails to open (#5172)
Addresses #5133. This was a bit messy but the simplest way I saw to
bubble up the error information. I added optional parameters to
BackendState that allow an error message to be intercepted and
displayed.


![image](https://github.com/user-attachments/assets/9945add9-b469-4329-ae15-34f979620c54)

I need some CSS help to make the div a reasonable relative width and for
the text to wrap.

---------

Co-authored-by: Brian Lambert <brianlambert@gmail.com>