Truncated long notebook outputs (#5117)
Addresses #5058 

This PR adds basic truncation to long text-based outputs in positron
notebooks. Previously we would just render the whole output inline which
could... really mess up navigating a notebook. Now we default to
truncation and respect the same options set for the vscode notebooks if
the user wants to switch to scrolling.


https://github.com/user-attachments/assets/97f1a1ab-ce2b-491a-8294-4a18c84d5f07

## Main change

### Handling Long Outputs:
* Introduced CSS styles for handling long outputs with scrolling and
truncation messages in `CellTextOutput.css`.
* Implemented logic in `CellTextOutput.tsx` to manage long outputs,
including truncation and scrolling behavior, and added a user interface
for changing output settings.

## Secondary changes
### Notebook Options Management:
* Added `useNotebookOptions` function to manage notebook options and
trigger rerendering when options change in
`NotebookInstanceProvider.tsx`.

### Notebook Instance Lifecycle:
* Updated `PositronNotebookInstance` to detach the view when an existing
instance is reused.
* Refactored the creation of `NotebookOptions` to use the instantiation
service, removing the dependency on `ICodeEditorService`.
[[1]](diffhunk://#diff-692a2d6ddf3fade96a06ef062de5b4282ba64e37294f3524f0b47877e4b3d3bcL219-R220)
[[2]](diffhunk://#diff-692a2d6ddf3fade96a06ef062de5b4282ba64e37294f3524f0b47877e4b3d3bcL248)
* Ensured `notebookOptions` is set to `undefined` when detaching the
view.

### Miscellaneous:
* Prevented cell deselection when clicking on links in
`NotebookCellWrapper.tsx`.
<!--
  Describe briefly what problem this pull request resolves, or what
  new feature it introduces. Include screenshots of any new or altered
  UI. Link to any GitHub issues but avoid "magic" keywords that will 
  automatically close the issue. If there are any details about your 
approach that are unintuitive or you want to draw attention to, please
  describe them here.
-->

### QA Notes

You should be able to run the following code in a notebook cell and get
truncation, the link should take you to the settings field for updating
the behavior.
```python
n = 100  # Change this value to control the number of lines
output = "\n".join(f"Line {i+1}" for i in range(n))
print(output)
```