check for appReadyMessage before previewing app
For apps that may output the server url before the app is ready, we need to check for the appReadyMessage before previewing the app.

Otherwise, the preview may fail because the app server is not ready to serve the app when the preview is requested.

In testing, fastapi apps were sometimes being previewed before the app was ready to serve the app.