Update how commands to source R files work (#5237)
Related to https://github.com/posit-dev/positron/pull/5200

I don't think we quite got there on making this behavior similar enough
to RStudio for folks. Notice that there are two commands:

![Screenshot 2024-10-31 at 1 25
21 PM](https://github.com/user-attachments/assets/c9f911a5-e41a-4868-bbcd-7473a782eae1)

In RStudio, this button "remembers" which one you chose last time so it
is understandable that folks (including me!) think it just works in a
certain way. However, it will be better to provide both. This will let
folks who are building a "Source on Save" keybinding or rule, as
outlined in
https://github.com/posit-dev/positron/pull/5200#issuecomment-2447294244,
to still have access to `r.sourceCurrentFile`.


### QA Notes

We can now either source with `echo = TRUE` or without an arg there, in
which case it will use `getOption("verbose")`. I set up the keybinding
in the RStudio Keymap as well.