Ignore branch when creating a version anchor (#5263)
We currently insist that you're on `main` when creating an anchor, to
avoid anchors being created that result in identical build numbers being
applied to different commits. However, this doesn't really apply now
that we have release branches with their own build sequence, so we need
to relax this restriction. This change just removes the superfluous
check.