Bump Ark to 0.1.150
# Release Notes

## New Features

- All top-level variables in R files are now indexed as workspace
symbols. You can now find assigned objects in the "Go to symbol in
Workspace feature" and use "Go to definition" to jump from a variable
name to its assignment (posit-dev/positron#5274). Repeated assignments
to the same name are not indexed yet but we plan to do so in the future.

- Completions for R functions and objects now display which package they
are from (posit-dev/positron#5225). And completions for package
namespaces are now indicated by a `::` suffix to make it clearer what is
being completed.

## Bug Fixes

- N/A