Bump Ark to 0.1.150 (#5275)
<!-- Thank you for submitting a pull request.
If this is your first pull request you can find information about
contributing here:
  * https://github.com/posit-dev/positron/blob/main/CONTRIBUTING.md

We recommend synchronizing your branch with the latest changes in the
main branch by either pulling or rebasing.
-->

<!--
  Describe briefly what problem this pull request resolves, or what
  new feature it introduces. Include screenshots of any new or altered
  UI. Link to any GitHub issues but avoid "magic" keywords that will 
  automatically close the issue. If there are any details about your 
approach that are unintuitive or you want to draw attention to, please
  describe them here.
-->

### Release Notes

#### New Features

- All top-level variables in R files are now indexed as workspace
symbols. You can now find assigned objects in the "Go to symbol in
Workspace feature" and use "Go to definition" to jump from a variable
name to its assignment (posit-dev/positron#5274). Repeated assignments
to the same name are not indexed yet but we plan to do so in the future.

- Completions for R functions and objects now display which package they
are from (posit-dev/positron#5225). And completions for package
namespaces are now indicated by a `::` suffix to make it clearer what is
being completed.

#### Bug Fixes

- N/A


### QA Notes

<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues..
-->