Smoke tests: Large python notebook (#5280)
This test uses this [large python
notebook](https://github.com/posit-dev/qa-example-content/blob/main/workspaces/large_py_notebook/spotify.ipynb)

It opens the notebook and runs it. Then it scrolls though about 2/3 of
the output collecting image refs to loosely ensure that the document
generated correctly.

It is a pretty heavy test. Not planning to run it on web and windows.
Tested on web and just running the notebook takes over a minute.

### QA Notes

All smoke tests should pass.