improve Run App url matching and fix preview timeout issue (#5336)
### Addresses
- https://github.com/posit-dev/positron/issues/5197
- https://github.com/posit-dev/positron/issues/5306

### Implementation Notes

#### URL Matching
- introduces `appUrlStrings` to the run and debug app options, which is
an optional array of `AppUrlString`s, which we will attempt to extract
the app url from
- `AppUrlString` is a string type that has the placeholder template
`{{APP_URL}}` in the string, which indicates the location of the app url
relative to the string
- adds the appropriate `appUrlStrings` for each framework we support
- expands on our url matching by:
- attempting to match url-like text where `{{APP_URL}}` is found in the
provided `appUrlStrings`
- if matching against `appUrlStrings` fails or no `appUrlStrings` are
found, we fallback to a more basic url match for strings that start with
http or https

#### Shell Integration Warning Message
This PR should also fix a timing issue where the `didPreviewUrlTimeout`
would time out before
`terminalOutputTimeout`, causing the shell integration warning message
to show. We now set `didPreviewUrlTimeout` to be 5 seconds longer than
`terminalOutputTimeout`, so that it doesn't timeout before the app
preview is done.

### QA Notes

This PR fixes Run App in Terminal url detection for non-local URLs. One
way to get a non-local url is by running the `shiny-py-example` in
Positron on Workbench (see
https://github.com/posit-dev/positron/issues/5197).

Other app types like Dash, Streamlit, Fastapi, Flask and Gradio should
continue to work on Desktop, Server Web and Positron on Workbench.

This PR should also fix the issue seen in #5306, which can be tested by
running the [Dash Py
Example](https://github.com/posit-dev/qa-example-content/tree/main/workspaces/dash-py-example)
several times to check that the issue does not occur.