add prerelease branches to the PR CI workflow (#5352)
Following
https://docs.github.com/en/actions/writing-workflows/workflow-syntax-for-github-actions#example-including-branches,
we can add the prerelease branch pattern `prerelease/**` so that the
usual PR CI is triggered when PRs are opened against prerelease
branches.

This is helpful for verifying backported fixes or fixes targeted for
prerelease branches.

I noticed this when I targeted my fix for #5197 for the prerelease
branch, merged it, but then ran into CI failures in my PR to bring in
the same fix to main. For the patch branch PR, I manually ran the Full
and Windows test suites, but missed the Python one.

### QA Notes

PRs against `prerelease` branches should now automatically kick off the
usual CI workflows that we see for PRs against `main`.