Fix vertical splitter resize behavior
The vertical splitter component would re-render itself while being resized. The renders would cause the React component to be re-created and the move event listeners to be detached from the original element the resize was happening on.

Event handling for the resizing operation is now done directly on the DOM elements and not as part of React state to preserve a link to the element that is not possible due to the nature of React component lifecycle.