add prerelease branched to merge workflow (#5388)
### Intent

Similar to https://github.com/posit-dev/positron/pull/5352, we should run full test suite when merging to prerelease branches.  This will help with catching issues when we aren't running the nightly full tests against the branch.

### QA Notes

Tests run when merging to prerelease branch. This will be hard to test until a change to a prerelease branch is made.