Fix restoration of R and Python sessions in server mode (for Workbench) (#5372)
This change fixes some issues around reconnecting to Python and R after
various actions that unload the UI (browser refresh, loading in a new
tab, closing a tab and coming back, reloading the desktop UI, etc).

The underlying issue is that we remember which Python and R sessions are
running by saving a list of active sessions into VS Code's workspace
storage. Unfortunately, workspace storage is not a suitable place to put
these things, because it is *too durable* -- it hangs around even when
Positron is closed -- and, in browser mode, it can also be shared among
workspaces when switching folders. Previous iterations of this change
tried to work around these problems by adding additional metadata or
clearing out the storage in some situations, but we were not able to
find a solution that covered all the edge cases.

To address this holistically we need a place to store the list of
sessions that is unique to the workspace _and_ the session -- that is,
it has the same lifetime as the server process that hosts Python and R.

This change adds that, in the form of a new service called an "Ephemeral
State Service". This service provides a way to store state that is
ephemeral (compared to the durable `IStorageService`) but needs to
persist across a user session, through browser refreshes and
disconnects/reconnects.

```mermaid
flowchart TD
subgraph "Positron Desktop"
f[Frontend Process] -- IPC Channel --> e[Electron Main]
e -- IPC Channel --> f
e --> m1[(In-Memory Ephemeral Store)]
end

subgraph "Positron Server"
b[Browser] -- Remote Agent Channel --> s[Server]
s -- Remote Agent Channel --> b
s --> m2[(In-Memory Ephemeral Store)]
end
```

We store the list of active sessions in this new service, so that the
session list has the same scope as the sessions themselves.

Addresses https://github.com/posit-dev/positron/issues/5267.

### QA Notes

This affects the way we remember which interpreter sessions are running
across all of Positron's various configurations. We don't have to test
every member of this matrix, but it wouldn't hurt to try to get a
representative sample:

Lifecycles:
- Close and reopen
- Reload page
- Exit / new session

Environments:
- Dev build
- Release build

Targets:
- Positron Desktop (Electron)
- Posit Workbench
- Remote SSH
- Positron Server (dev only)

The main class of errors will be either (a) attempts to reconnect to
sessions that don't exist, and (b) failure to reconnect to sessions that
_do_ exist. In either case the most helpful logs are from the Developer
Tools as that's where most of this code logs.